Android application
PlantID is an android application, which takes the picture of the leaf
against white background. Picture gets saved in the phone memory and as
well as in the image gallery in the app. We can compare the picture of
the leaf with the existing images or details of the leaves in the
database. If the same leaf exist in the database then details of the
leaf will pop up on the screen otherwise we can add the details of the
leaf manually. We can also perform delete, refresh, zoom in , zoom out
actions on the pictures. Still working on the visual recognition tool
for the application.